Move getModuleName into fs-utils, add tests
