+++ packages/hekla-core/src/Analyzer.js
@@ -5,6 +5,9 @@ const {
 
 const Module = require('./Module');
 const {
+  getModuleName
+} = require('./utils/fs-utils');
+const {
   parseAST,
   parseHTML,
   ASTWrapper,
@@ -38,10 +41,6 @@ module.exports = class Analyzer {
     this.fs = fs;
   }
 
-  getModuleName(resource) {
-    return getModuleName(resource, this.rootPath);
-  }
-
   getAnalysis() {
     const analysis = {
       modules: this.modules.map(module => module.serialize())
@@ -50,7 +49,7 @@ module.exports = class Analyzer {
   }
 
   createModule(resource) {
-    const moduleName = this.getModuleName(resource);
+    const moduleName = getModuleName(resource, this.rootPath);
     return new Module(moduleName, resource);
   }
 
@@ -111,14 +110,4 @@ function readFile(fs, filename) {
       }
     });
   });
-}
-
-function getModuleName(resource, rootPath) {
-  let fullPath = resource;
-  if (fullPath.indexOf('!') !== -1) {
-    const pieces = resource.split('!');
-    fullPath = pieces[pieces.length - 1];
-  }
-  const projectPath = fullPath.replace(rootPath, '');
-  return `.${projectPath}`;
 }
